Refactor to use components and support IPC; fixes #24; sadly one file has many whitespace changes due to autoformat, but decided to bite the bullet
